requirements/dev.txt
-r ./base.txt

pytest==6.2.2
pytest-django==4.1.0
factory_boy==3.2.0
black==20.8b1
isort==5.7.0
